テストの追加

+++ task-backend/db/migrate/20231012020047_create_sections.rb
@@ -2,7 +2,7 @@ class CreateSections < ActiveRecord::Migration[6.1]
   def change
     create_table :sections do |t|
       t.string :title
-      t.integer :position
+      t.float :position
 
       t.timestamps
     end
